requirements.txt
-e https://github.com/gar1t/pydiscourse.git#egg=pydiscourse
pyyaml
requests
